cleanup imports and unused vars
