+++ examples/frontend/src/components/DropdownMenu.tsx
@@ -1,11 +1,4 @@
 import {
-    Fragment,
-    useState,
-    Dispatch,
-    SetStateAction
-} from "react";
-
-import {
     Listbox,
     Transition,
 } from "@headlessui/react";
