+++ examples/frontend/src/pages/bridge/BridgePage.tsx
@@ -1,48 +1,19 @@
-import {useContext, useEffect, useState} from "react";
+import {useState} from "react";
 
-import {BridgeDirections} from "./Directions";
-
-import NetworkDropdown from "./components/NetworkDropdown";
-import TokenDropdown from "./components/TokenDropdown";
-import BridgeEstimateSection from "./components/BridgeEstimateSection";
-import AmountFromDropdown, {AmountDropdownItem} from "./components/AmountFromDropdown";
 import ApproveButon from "./components/ApproveButon";
 
-import {TokenMenuContext,   TokenMenuContextProvider}   from "./contexts/TokenMenuContext";
-import {NetworkMenuContext, NetworkMenuContextProvider} from "./contexts/NetworkMenuContext";
-
-import type {Token} from "@synapseprotocol/sdk";
-import {
-    Networks,
-    detailedTokenSwapMap,
-    supportedNetworks,
-} from "@synapseprotocol/sdk";
+import {TokenMenuContextProvider}   from "./contexts/TokenMenuContext";
+import {NetworkMenuContextProvider} from "./contexts/NetworkMenuContext";
 
 import {Grid} from "../../components/Grid";
 
 import {BigNumber} from "ethers";
-
-import {DropdownItem} from "../../components/DropdownMenu";
 
 import SourceGrid, {AMOUNTS_FROM_OPTIONS} from "./SourceGrid";
 import DestinationGrid from "./DestinationGrid";
 import BridgeButton from "./components/BridgeButton";
 
 function BridgePageContent(props: {className?: string}) {
-    const {
-        selectedNetworkFrom,
-        selectedNetworkTo
-    } = useContext(NetworkMenuContext);
-
-    const {
-        selectedTokenFrom,
-        selectedTokenTo
-    } = useContext(TokenMenuContext);
-
-    // useEffect(() => {
-    //     console.log({selectedNetworkFrom});
-    // }, [selectedNetworkFrom])
-
     const [amountFrom, setAmountFrom] = useState(AMOUNTS_FROM_OPTIONS[3]);
 
     const [approved, setApproved] = useState<boolean>(true);
