--- examples/frontend/src/pages/bridge/components/ActionButton.tsx
@@ -1,21 +1,0 @@
-export type OnClickFunction = (...args: any[]) => (void | any);
-
-export interface ButtonProps {
-    text:      string,
-    onClick?:  OnClickFunction,
-    disabled?: boolean,
-}
-
-export const emptyOnClick: OnClickFunction = () => {}
-
-export const ActionButton = ({text, onClick, disabled=false}: ButtonProps) => (
-    <div>
-        <button
-            className={"w-full bg-blue-500 hover:bg-blue-700 text-white text-sm font-bold py-2 px-4 rounded-full"}
-            onClick={onClick ?? emptyOnClick}
-            disabled={disabled}
-        >
-            {text}
-        </button>
-    </div>
-)
