+++ examples/frontend/src/pages/bridge/hooks/useActionButtonOnClick.tsx
@@ -1,7 +1,6 @@
 import {ethers} from "ethers";
 import {SetStateFunction} from "../../../utils";
-import {ButtonProps} from "../components/ActionButton";
-
+import {ButtonProps} from "../../../components/Button";
 
 
 export function useActionButtonOnClick(
